In example code, replace deprecated function
`getAllUserHoldings` is deprecated, so it should not be used in the example code.

Use `listUserAccounts` in the example instead. This function is similar to `getAllUserHoldings` in that it only requires userId and userSecret. So it makes sense that this would be chronologically first in a series of other API calls.

The example could be expanded further by then using the list of accounts to make additional API calls, but for now, we have a minimal example that does not use any deprecated functions.